+++ NEWS.md
@@ -1,5 +1,6 @@
 # cfb4th (development version)
 
+* Re-categorized some plays as unknown (i.e., `NA`) `go`: Penalties and Timeouts
 * Fixed bug with 4th down plays inside own 10 not simulating failed go for it plays correctly (0.1.0.9001)
 
 # cfb4th 0.1.0